// src/app/admin/models/blogs.model.ts
export class Blogs {
    id?: string | number;
    title?: string | undefined;
    videolink?: string | undefined;
    question?: string | undefined;
    program? = '';
    output?: string | undefined;
    explinaion?:'';
    courseKeyword?: string | undefined;
    button_name?: string | undefined;
    publish?: boolean | false;
}